split utils functions into docker functions & file system funtions
