+++ utils/utils_test.go
@@ -8,6 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
+
 package utils
 
 import (
@@ -85,13 +86,13 @@ func TestCreateTempFile(t *testing.T) {
 
 func TestWriteToComposeFile(t *testing.T) {
 	os.Create("TestFile.yaml")
-	got := WriteToComposeFile("TestFile.yaml")
+	got := WriteToComposeFile("TestFile.yaml", false)
 	assert.Equal(t, got, true, "should return true: should write data to a temp file")
 	os.Remove("TestFile.yaml")
 }
 
 func TestWriteToComposeFileFail(t *testing.T) {
-	writeToFile := WriteToComposeFile("")
+	writeToFile := WriteToComposeFile("", false)
 	assert.Equal(t, writeToFile, false, "should return false: should fail to write data")
 }
